+++ CHANGELOG.md
@@ -23,6 +23,13 @@ and this project does not adhere to [Semantic Versioning](http://semver.org/).
   Use the auxiliary functions `iterateOverLocalIndexSet` and
   `addToGlobalMatrix` to conveniently handle index sets from both
   interfaces.
+* Add a Bernstein polynomial basis under the name
+  `BernsteinPkLocalFiniteElement` with corresponding global bases
+  `BernsteinBasis`, `BernsteinDGBasis` and `BernsteinDGRefinedDGNodalBasis`.
+  They can be used as a more stable replacement for the standard
+  Lagrange basis.
+  What is still missing is a continuous Bernstein basis with hanging
+  node constraints.
 * New functions to save grid functions before adaptive refinement and to
   restore them to the refined grid, namely `attachDataToGrid` and
   `restoreDataToRefinedGrid`.